모든 파일 타입 수정(Integer)

+++ backend/src/main/java/com/example/backend/product/dto/CartDeleteRequest.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class CartDeleteRequest {
-    private Long cartId;
+    private Integer cartId;
 }
