+++ backend/src/main/java/com/example/backend/product/repository/CartRepository.java
@@ -8,14 +8,16 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CartRepository extends JpaRepository<Cart, Long> {
-    List<Cart> findByMemberId(Long memberId);
+public interface CartRepository extends JpaRepository<Cart, Integer> {
+    List<Cart> findByMemberId(Integer memberId);
 
     List<Cart> findByMemberAndProductAndOption(Member member, Product product, ProductOption option);
 
 //    void deleteByMemberIdAndProductIdAndOptionId(Long memberId, Long productId, Long optionId);
 
-    void deleteByMemberId(Long memberId);
+    void deleteByMemberId(Integer memberId);
+
+//    List<Cart> findByMemberId(Integer memberId);
 
 //    List<Cart> findByUser(User user)
 }
